feat: move gameBoard to GameTable
