+++ src/components/GameTable/style.ts
@@ -1,5 +1,16 @@
 import { CSSProperties } from 'react';
 
+export const gameBoardStyle: CSSProperties = {
+    margin: ' 5px auto 10px',
+    flex: 1,
+    width: '100%',
+    maxWidth: '500px',
+    display: 'flex',
+    flexDirection: 'column',
+    position: 'relative',
+    paddingTop: 0,
+};
+
 export const gameTableContainerStyle: CSSProperties = {
     height: '100%',
     paddingTop: '100%',
